Moving the 'DataframesDBKitInterface' class to the common folder
- Replaced the class name from 'ExtratoDBKitInterface' to 'DataframesDBKitInterface'
- Moved the 'DataframesDBKitInterface' class from the 'extrato_dataframes_kit.py' to the 'common/dataframes_kit.py' file
